+++ packages/location_platform_interface/test/method_channel_location_test.dart
@@ -67,6 +67,7 @@ void main() {
           'interval': 1000,
           'distanceFilter': 0,
           'pausesLocationUpdatesAutomatically': true,
+          'backgroundInterval': null,
         },
       ),
     ]);